requirements.txt
amqp==5.2.0
asgiref==3.8.1
celery[redis]==5.2.7
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==3.2.3
django-model-utils==4.5.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
importlib-metadata==4.13.0
kombu==5.3.7
mccabe==0.7.0
prompt_toolkit==3.0.45
psycopg2==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
zipp==3.19.1
